USB configuration #3 #2
Brought USB configuration into the program. We added a button at A0 that
controls what mode the system is in. If the button is pressed (grounded)
then the system boots into USB mode. Otherwise, it boots into flight
control mode.

The USB device and serial device is instantiated no matter what, but if
it sits idle this doesn't impact performance and makes it significantly
easier from an architecture standpoint.