Logging changes #5
Refactored the logger a little bit. Logging is implemented (and tested...ish). Moved management of the W25Q flash into the logging module itself so that writes are more managed.

Created FIFO queue system for logging. Logs are written to a bounded region which is set at initializtion. The first block is reserved. The size of the region is written to the first page of the first block so that we know how large the region is. Each new sector is erased before being written.