New shared library
The way I'm thinking we do this is that all of the application logic
will be in a library that can be shared across multiple different
hardware applications. Individual BBU tasks will be in "examples" that
are able to be developed independently. So, starting with the pressure
sensor, we have an example code that called bbu-bmp388.rs that will
exercise all possible functions of the pressure sensor. Successful BBU
of the pressure sensor will be when this example code runs successfully
and exercises everything we need.

Then, we can copy from the example code when we actually want to integrate
the pressure sensor into the application.